format price & add more seed

+++ database/migrations/2021_08_25_022213_create_sell_summaries_table.php
@@ -22,7 +22,7 @@ class CreateSellSummariesTable extends Migration
             $table->dateTime('created_date');
             $table->dateTime('last_update')->nullable();
             $table->integer('price_total');
-            $table->float('discount_total');
+            $table->integer('discount_total');
             $table->integer('total');
             $table->timestamps();
         });
